Make mandala-page folder.
