+++ lib/pages/mandala-page/mandala-design.avsc.json
